+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - Fixed bug in `get_authenticated_session` that occasionally caused tests to fail with 'Connection reset by peer' errors.
 
 
+## Changed
+- `AudioPrompt` has had the default play control label renamed from 'Play from start' to just 'Play'.
+
 # [12.1.1](https://gitlab.com/PsyNetDev/PsyNet/-/releases/v12.1.1) Release - 2025-07-15
 
 ## Fixed